correction de type du variables

+++ src/bookstore/ImpUtility.java
@@ -6,13 +6,13 @@ public class ImpUtility {
 
 	public static void main(String[] args) {
 		int qte;
-		int unitPrice;
+		double unitPrice;
 		double prixTotal ;
 		Scanner sq = new Scanner(System.in);
 		System.out.println("enter the quantity");
 		qte = sq.nextInt();
 		System.out.println("enter the unit price");
-		unitPrice = sq.nextInt();
+		unitPrice = sq.nextDouble();
 		Utility u = new Utility();
 		prixTotal = u.calculateTotalPrice(qte, unitPrice) ;
 		System.out.println("Total price = " + prixTotal );
